+++ src/extensions/constants.py
@@ -1,0 +1,5 @@
+FULL_CHAR = "█"
+HALF_CHAR = "▒"
+TOTAL_PROGRESS = "[{current}/{total}]"  # [15/100]
+PERCENT_PROGRESS = "[{percent:.2f}%]"  # [15.00%]
+RAW_PROGRESS_BAR_LENGTH = 64
